requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pandas==2.2.3
openpyxl==3.1.5
python-multipart==0.0.12
python-dotenv==1.0.1
